Only include vrrp header files when building with vrrp and also for check
Make sure vrrp header files are only included if building with vrrp (i.e.
without --disable-vrrp), and likewise only include check header files if
compiling with LVS support (i.e. without --disable-lvs).
