+++ keepalived/include/check_api.h
@@ -29,16 +29,7 @@
 /* local includes */
 #include "check_data.h"
 #include "scheduler.h"
-
-/* connection options structure definition */
-typedef struct _conn_opts {
-	struct sockaddr_storage		dst;
-	struct sockaddr_storage		bindto;
-	unsigned int			connection_to; /* connection time-out */
-#ifdef _WITH_SO_MARK_
-	unsigned int			fwmark; /* to mark packets going out of the socket using SO_MARK */
-#endif
-} conn_opts_t;
+#include "layer4.h"
 
 /* Checkers structure definition */
 typedef struct _checker {
