+++ keepalived/include/ipvswrapper.h
@@ -41,14 +41,16 @@
 #ifdef _WITH_LVS_
   #include "libipvs.h"
   #include "ip_vs.h"
-#ifdef _WITH_LVS_
+#ifdef _WITH_VRRP_
   #include "vrrp.h"
 #endif
 #endif
 
 /* locale includes */
 #include "scheduler.h"
+#ifdef _WITH_LVS_
 #include "check_data.h"
+#endif
 
 #define IPVS_ERROR	0
 #define IPVS_SUCCESS	1
@@ -84,7 +86,7 @@ do {						\
 		UNSET_ALIVE((V));		\
 } while (0)
 
-#ifdef _WITH_LVS_
+#if defined _WITH_VRRP_ && defined _WITH_LVS_
 struct lvs_syncd_config {
 	char				*ifname;	/* handle LVS sync daemon state using this */
 	vrrp_t				*vrrp;		/* instance FSM & running on specific interface */
@@ -108,7 +110,7 @@ extern virtual_server_group_t *ipvs_get_group_by_name(char *, list);
 extern void ipvs_group_sync_entry(virtual_server_t *vs, virtual_server_group_entry_t *vsge);
 extern void ipvs_group_remove_entry(virtual_server_t *, virtual_server_group_entry_t *);
 extern int ipvs_cmd(int, virtual_server_t *, real_server_t *);
-#ifdef _WITH_LVS_
+#ifdef _WITH_VRRP_
 extern void ipvs_syncd_cmd(int, const struct lvs_syncd_config *, int, bool, bool);
 extern void ipvs_syncd_master(const struct lvs_syncd_config *);
 extern void ipvs_syncd_backup(const struct lvs_syncd_config *);
