+++ keepalived/include/layer4.h
@@ -34,7 +34,6 @@
 
 /* local includes */
 #include "scheduler.h"
-#include "check_api.h"
 
 enum connect_result {
 	connect_error,
@@ -43,9 +42,21 @@ enum connect_result {
 	connect_success
 };
 
+/* connection options structure definition */
+typedef struct _conn_opts {
+	struct sockaddr_storage		dst;
+	struct sockaddr_storage		bindto;
+	unsigned int			connection_to; /* connection time-out */
+#ifdef _WITH_SO_MARK_
+	unsigned int			fwmark; /* to mark packets going out of the socket using SO_MARK */
+#endif
+} conn_opts_t;
+
 /* Prototypes defs */
+#ifdef _WITH_LVS_
 extern enum connect_result
  socket_bind_connect(int, conn_opts_t *);
+#endif
 
 extern enum connect_result
  socket_connect(int, struct sockaddr_storage *);
@@ -53,17 +64,21 @@ extern enum connect_result
 extern enum connect_result
  socket_state(thread_t *, int (*func) (thread_t *));
 
+#ifdef _WITH_LVS_
 extern int
  socket_connection_state(int, enum connect_result
 		      , thread_t *, int (*func) (thread_t *)
 		      , unsigned long);
+#endif
 
 /* Backward compatibility */
+#ifdef _WITH_LVS_
 static inline enum connect_result
 tcp_bind_connect(int fd, conn_opts_t *co)
 {
 	return socket_bind_connect(fd, co);
 }
+#endif
 
 static inline enum connect_result
 tcp_connect(int fd, struct sockaddr_storage *addr)
@@ -77,10 +92,13 @@ tcp_socket_state(thread_t * thread, int (*func) (thread_t *))
 	return socket_state(thread, func);
 }
 
+#ifdef _WITH_LVS_
 static inline int
 tcp_connection_state(int fd, enum connect_result status, thread_t * thread,
              int (*func) (thread_t *), unsigned long timeout)
 {
 	return socket_connection_state(fd, status, thread, func, timeout);
 }
+#endif
+
 #endif
