+++ keepalived/include/main.h
@@ -35,8 +35,12 @@
 #include "pidfile.h"
 #include "scheduler.h"
 #include "parser.h"
+#ifdef _WITH_VRRP_
 #include "vrrp_daemon.h"
+#endif
+#ifdef _WITH_LVS_
 #include "check_daemon.h"
+#endif
 #include "global_data.h"
 
 /* State flags */
