+++ keepalived/include/smtp.h
@@ -29,7 +29,12 @@
 /* local includes */
 #include "scheduler.h"
 #include "layer4.h"
+#ifdef _WITH_LVS_
+#include "check_data.h"
+#endif
+#ifdef _WITH_VRRP_
 #include "vrrp.h"
+#endif
 
 /* global defs */
 #define SMTP_PORT_STR		"25"
@@ -88,6 +93,14 @@ typedef struct _smtp {
 #define SMTP_QUIT_CMD    "QUIT\r\n"
 
 #define FMT_SMTP_HOST()	inet_sockaddrtopair(&global_data->smtp_server)
+
+#ifndef _WITH_LVS_
+typedef void real_server_t;
+#endif
+#ifndef _WITH_VRRP_
+typedef void vrrp_t;
+typedef void vrrp_sgroup_t;
+#endif
 
 /* Prototypes defs */
 extern void smtp_alert(real_server_t *, vrrp_t *, vrrp_sgroup_t *,
