+++ keepalived/include/vrrp_if.h
@@ -49,9 +49,6 @@
 #define LB_MII     0x2
 #define LB_ETHTOOL 0x4
 
-/* Default values */
-#define IF_DEFAULT_BUFSIZE	(65*1024)
-
 /* I don't know what the correct type is. 
  * The kernel has ifindex in the range [1, INT_MAX], but IFLA_LINK is defined
  * to be __u32. See dev_new_index() in net/core/dev.c and net/core/rtnetlink.c.
